+++ example/lib/main.dart
@@ -5,8 +5,16 @@ void main() {
   runApp(const MyApp());
 }
 
-class MyApp extends StatelessWidget {
+class MyApp extends StatefulWidget {
   const MyApp({super.key});
+
+  @override
+  State<MyApp> createState() => _MyAppState();
+}
+
+class _MyAppState extends State<MyApp> {
+  bool colorChanged = false;
+  Color color = Colors.deepPurple;
 
   @override
   Widget build(BuildContext context) {
@@ -14,27 +22,35 @@ class MyApp extends StatelessWidget {
       title: 'Flutter Sliding Toast Demo',
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurpleAccent),
+        colorScheme: ColorScheme.fromSeed(
+          seedColor: color,
+        ),
+        appBarTheme: AppBarTheme(
+          backgroundColor: color,
+          foregroundColor: Colors.white,
+        ),
         useMaterial3: true,
       ),
-      home: const MyHomePage(),
+      home: MyHomePage(changeColor: changeColor),
     );
   }
+
+  changeColor() => setState(() {
+        colorChanged ? color = Colors.deepPurple : color = Colors.green;
+        colorChanged = !colorChanged;
+      });
 }
 
 class MyHomePage extends StatelessWidget {
-  const MyHomePage({super.key});
+  final VoidCallback changeColor;
+  const MyHomePage({super.key, required this.changeColor});
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         iconTheme: const IconThemeData(color: Colors.white),
-        backgroundColor: Colors.deepPurpleAccent,
-        title: const Text(
-          "Sliding Toast",
-          style: TextStyle(color: Colors.white),
-        ),
+        title: const Text("Sliding Toast"),
         centerTitle: true,
       ),
       body: SafeArea(
@@ -47,11 +63,10 @@ class MyHomePage extends StatelessWidget {
                 SlidingToast.show(
                   context,
                   title: const Text(
-                    "Hi there! I'm a simple toast 😎. Dismiss me by sliding downward.",
-                    style: TextStyle(),
+                    "Hi there! I'm a simple toast 😎."
+                    " Dismiss me by sliding downward.",
                   ),
-                  trailing:
-                      const Icon(Icons.person, color: Colors.deepPurpleAccent),
+                  trailing: const TrailingWidget(),
                   toastSetting: const ToastSetting(
                     animationDuration: Duration(seconds: 1),
                     displayDuration: Duration(seconds: 2),
@@ -60,7 +75,10 @@ class MyHomePage extends StatelessWidget {
                   ),
                 );
               },
-              child: const Text("Toast from bottom center"),
+              child: const Text(
+                "Toast from bottom center",
+                style: TextStyle(fontSize: 16),
+              ),
             ),
             const SizedBox(height: 30),
             ElevatedButton(
@@ -68,10 +86,10 @@ class MyHomePage extends StatelessWidget {
                 SlidingToast.show(
                   context,
                   title: const Text(
-                    "Hi there! I'm a glassy toast 😎. Dismiss me by sliding upward.",
-                    style: TextStyle(),
+                    "Hi there! I'm a glassy toast 🦄."
+                    "Dismiss me by sliding upward.",
                   ),
-                  trailing: const Icon(Icons.person),
+                  trailing: const TrailingWidget(),
                   toastSetting: const ToastSetting(
                     animationDuration: Duration(seconds: 1),
                     displayDuration: Duration(seconds: 2),
@@ -85,7 +103,10 @@ class MyHomePage extends StatelessWidget {
                   ),
                 );
               },
-              child: const Text("Glassy toast from top center"),
+              child: const Text(
+                "Glassy toast from top center",
+                style: TextStyle(fontSize: 16),
+              ),
             ),
             const SizedBox(height: 30),
             ElevatedButton(
@@ -93,8 +114,8 @@ class MyHomePage extends StatelessWidget {
                 SlidingToast.showError(
                   context,
                   title: const Text(
-                    "Hi there! I'm a error toast 😈. Dismiss me by sliding horizontally.",
-                    style: TextStyle(),
+                    "Hi there! I'm a error toast 😈. "
+                    "Dismiss me by sliding horizontally.",
                   ),
                   toastSetting: const ToastSetting(
                     toastStartPosition: ToastPosition.left,
@@ -103,7 +124,10 @@ class MyHomePage extends StatelessWidget {
                   ),
                 );
               },
-              child: const Text("Error toast from bottom left"),
+              child: const Text(
+                "Error toast from bottom left",
+                style: TextStyle(fontSize: 16),
+              ),
             ),
             const SizedBox(height: 30),
             ElevatedButton(
@@ -111,8 +135,7 @@ class MyHomePage extends StatelessWidget {
                 SlidingToast.showSuccess(
                   context,
                   title: const Text(
-                    "Hi there! I'm a success toast 😎",
-                    style: TextStyle(),
+                    "Hi there! I'm a success toast 🦁",
                   ),
                   toastSetting: const ToastSetting(
                     toastStartPosition: ToastPosition.right,
@@ -120,7 +143,10 @@ class MyHomePage extends StatelessWidget {
                   ),
                 );
               },
-              child: const Text("Success toast from Top Right"),
+              child: const Text(
+                "Success toast from Top Right",
+                style: TextStyle(fontSize: 16),
+              ),
             ),
             const SizedBox(height: 30),
             ElevatedButton(
@@ -128,7 +154,7 @@ class MyHomePage extends StatelessWidget {
                 SlidingToast.show(
                   context,
                   title: const Text(
-                    "Hi there! I'm modified toast 😎 with only title widget "
+                    "Hi there! I'm modified toast 🐷 with only title widget "
                     "and display duration of 4 seconds. "
                     " Check out the code for more details.",
                     textAlign: TextAlign.center,
@@ -144,10 +170,10 @@ class MyHomePage extends StatelessWidget {
                     toastAlignment: Alignment.center,
                   ),
                   toastStyle: ToastStyle(
-                    backgroundColor: Colors.deepPurpleAccent,
+                    backgroundColor: Colors.blue,
                     boxShadow: [
                       BoxShadow(
-                        color: Colors.deepPurpleAccent.withOpacity(.2),
+                        color: Colors.blue.withOpacity(.2),
                         blurRadius: 5,
                         spreadRadius: 2,
                       )
@@ -155,11 +181,50 @@ class MyHomePage extends StatelessWidget {
                   ),
                 );
               },
-              child: const Text("Modified toast from center left"),
+              child: const Text(
+                "Modified toast from center left",
+                style: TextStyle(fontSize: 16),
+              ),
+            ),
+            const SizedBox(height: 30),
+            ElevatedButton(
+              onPressed: () {
+                SlidingToast.show(
+                  context,
+                  title: const Text(
+                    "Hi there! I'm a callback toast 🐺."
+                    "Tap me to change theme color.",
+                  ),
+                  trailing: const TrailingWidget(),
+                  toastSetting: const ToastSetting(
+                    animationDuration: Duration(seconds: 1),
+                    displayDuration: Duration(seconds: 4),
+                    toastStartPosition: ToastPosition.bottom,
+                    toastAlignment: Alignment.bottomCenter,
+                  ),
+                  onTapped: changeColor,
+                );
+              },
+              child: const Text(
+                "Callback toast",
+                style: TextStyle(fontSize: 16),
+              ),
             ),
           ],
         ),
       ),
+    );
+  }
+}
+
+class TrailingWidget extends StatelessWidget {
+  const TrailingWidget({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Icon(
+      Icons.person,
+      color: Theme.of(context).colorScheme.primary,
     );
   }
 }