+++ lib/flutter_sliding_toast.dart
@@ -3,3 +3,4 @@ library flutter_sliding_toast;
 export 'package:flutter_sliding_toast/src/sliding_toast.dart';
 export 'package:flutter_sliding_toast/src/toast_setting.dart';
 export 'package:flutter_sliding_toast/src/toast_style.dart';
+export 'package:flutter_sliding_toast/src/enums/toast_position.dart';