+++ lib/src/toast_setting.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 
+import 'enums/toast_position.dart';
+
 /// The setting for the sliding animation
 @immutable
 class ToastSetting {
@@ -90,53 +92,5 @@ class ToastSetting {
       toastAlignment: toastAlignment ?? this.toastAlignment,
       padding: padding ?? this.padding,
     );
-  }
-}
-
-/// The toast will start sliding from this direction
-enum ToastPosition {
-  top,
-  right,
-  bottom,
-  left;
-
-  /// Tween with start and end offsets with according to the toast start position
-  Tween<Offset> tween() {
-    switch (this) {
-      case ToastPosition.top:
-        return Tween(
-          begin: const Offset(0, -1),
-          end: const Offset(0, 0),
-        );
-      case ToastPosition.right:
-        return Tween(
-          begin: const Offset(1, 0),
-          end: const Offset(0, 0),
-        );
-      case ToastPosition.bottom:
-        return Tween(
-          begin: const Offset(0, 1),
-          end: const Offset(0, 0),
-        );
-      case ToastPosition.left:
-        return Tween(
-          begin: const Offset(-1, 0),
-          end: const Offset(0, 0),
-        );
-    }
-  }
-
-  /// Dismiss direction according to the toast start position
-  DismissDirection dismissDirection() {
-    switch (this) {
-      case ToastPosition.top:
-        return DismissDirection.up;
-      case ToastPosition.right:
-        return DismissDirection.horizontal;
-      case ToastPosition.bottom:
-        return DismissDirection.down;
-      case ToastPosition.left:
-        return DismissDirection.horizontal;
-    }
   }
 }